python/requirements.txt
click==8.1.2
ezdxf==0.17.2
numpy==1.21.5
opencv-python==4.5.5.64
pybind11==2.9.2
PyYAML==6.0
tqdm==4.63.1
typer==0.4.1
zipp==3.7.0
